README appears to not use correct identifiers
not 100% on this since I only read the readme, but it appears that
`data-wt="alert"` defines the identifier as `alert`. But in further
sections of the toy example, `alert` appears to be switched out for
`popup` in one case and `AlertButton` in another.

Looks like a few README edit committs may have jumbled a few versions of
the opening example together.